backend/generate-pdf/requirements.txt
reportlab>=4.0.0
psycopg2-binary>=2.9.0
boto3>=1.28.0
